added notfound exception

+++ src/features/api/super-admin/quiz/handlers/deleteQuestion.handler.ts
@@ -1,5 +1,6 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { QuizRepository } from '../quiz.repository';
+import { NotFoundException } from '@nestjs/common';
 
 export class DeleteQuestionCommand {
   constructor(public id: string) {}
@@ -11,7 +12,9 @@ export class DeleteQuestionHandler
 {
   constructor(private repo: QuizRepository) {}
 
-  execute(command: DeleteQuestionCommand): Promise<any> {
+  async execute(command: DeleteQuestionCommand): Promise<any> {
+    const question = await this.repo.getOne(command.id);
+    if (!question) throw new NotFoundException();
     return this.repo.delete(command.id);
   }
 }
